requirements.txt
streamlit
groq
langchain
langchain-groq
scipy
numpy
python-dotenv
